Use socket activation using the systemd library
This works fine when there are no errors. However, when an error occurs
during the first connection, the socket will enter the failed state, and
subsequent connections will not even trigger a service start. The same
happens when we do restarts on failure, it just tries a few times after
the first connection, and then it gives up.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>