Wait for xapi to be up before trying to log in
Previously when the server tried to login at startup, it failed
immediately because xapi wasn't ready yet, even though systemd starts
xapi-nbd after xapi. The proper long term fix will be for xapi to notify
systemd when its startup sequence has completed, and when that is done,
this temporary fix can be removed.

Waiting for xapi's Unix domain socket to be up does not work, there
would still be transient failures because apparently xapi is not
completely ready when the Unix domain socket is created.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>